♻️ Move SymptomsPage to its own package
Deletes Symptoms.kt from the navigation package
