+++ app/src/main/kotlin/br/com/colman/petals/navigation/Page.kt
@@ -45,6 +45,7 @@ import br.com.colman.petals.hittimer.ComposeHitTimer
 import br.com.colman.petals.navigation.Page.Usage
 import br.com.colman.petals.settings.SettingsView
 import br.com.colman.petals.statistics.StatisticsPage
+import br.com.colman.petals.withdrawal.Symptoms
 import org.koin.androidx.compose.get
 
 enum class Page(
