+++ app/src/main/kotlin/br/com/colman/petals/withdrawal/SymptomsPage.kt
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package br.com.colman.petals.navigation
+package br.com.colman.petals.withdrawal
 
 import androidx.compose.foundation.layout.Arrangement.Absolute.spacedBy
 import androidx.compose.foundation.layout.Box
